Fold the sound normalizer into analyze/range_restriction.py (one module)
The previous commit left two range_restriction files in analyze/ (the check and
a separate _normalize module). Merge them: the sound normalization
(phantom_paths, PhantomProjection, ValueSplit, RangeRestrictionNormalizer) now
lives in dyna/analyze/range_restriction.py alongside the refined check, a single
standalone module next to abbreviate. Deletes range_restriction_normalize.py.

Suite: 336 passed. Doc module map updated.
